Remove numpy dependency
