+++ torchmetrics/utilities/data.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 from typing import Any, Callable, List, Mapping, Optional, Sequence, Union
 
-import numpy as np
 import torch
 from torch import Tensor, tensor
 
@@ -202,23 +201,22 @@ def apply_to_collection(
     return data
 
 
-def get_group_indexes(indexes: Union[Tensor, np.ndarray]) -> List[Union[Tensor, np.ndarray]]:
+def get_group_indexes(indexes: Tensor) -> List[Tensor]:
     """
-    Given an integer `torch.Tensor` or `np.ndarray` `indexes`, return a `torch.Tensor` or `np.ndarray` of indexes for
+    Given an integer `torch.Tensor` `indexes`, return a `torch.Tensor` of indexes for
     each different value in `indexes`.
 
     Args:
-        indexes: a `torch.Tensor` or `np.ndarray` of integers
+        indexes: a `torch.Tensor`
 
     Return:
-        A list of integer `torch.Tensor`s or `np.ndarray`s
+        A list of integer `torch.Tensor`s
 
     Example:
         >>> indexes = torch.tensor([0, 0, 0, 1, 1, 1, 1])
         >>> get_group_indexes(indexes)
         [tensor([0, 1, 2]), tensor([3, 4, 5, 6])]
     """
-    structure, dtype = (tensor, torch.long) if isinstance(indexes, Tensor) else (np.array, np.int64)
 
     res = dict()
     for i, _id in enumerate(indexes):
@@ -228,4 +226,4 @@ def get_group_indexes(indexes: Union[Tensor, np.ndarray]) -> List[Union[Tensor, 
         else:
             res[_id] = [i]
 
-    return [structure(x, dtype=dtype) for x in res.values()]
+    return [tensor(x, dtype=torch.long) for x in res.values()]
